feat: create useChatForm to handle send text or donation message
